+++ src/aabb.hpp
@@ -19,8 +19,8 @@ struct AABB{
 };
 
 uint8_t createAABB(Stage& stage, AABB aabb);
-size_t deleteAABBByIdx(Stage& stage, int aabb_idx);
-size_t deleteAABBById(Stage& stage, uint8_t aabb_id);
+int deleteAABBByIdx(Stage& stage, int aabb_idx);
+int deleteAABBById(Stage& stage, uint8_t aabb_id);
 
 // TODO: program
